Move initializer to device namespace
